refactor(gui): isolate embedded web contracts from main
