+++ crates/gwt/src/main.rs
@@ -13,7 +13,7 @@ use axum::{
         ws::{Message, WebSocket, WebSocketUpgrade},
         State,
     },
-    response::{Html, IntoResponse},
+    response::IntoResponse,
     routing::get,
     Router,
 };
@@ -41,6 +41,8 @@ use tokio::{
 };
 use uuid::Uuid;
 use wry::WebViewBuilder;
+
+mod embedded_web;
 
 type ClientId = String;
 const DEFAULT_NEW_BRANCH_BASE_BRANCH: &str = "develop";
@@ -1827,95 +1829,6 @@ mod tests {
             dunce::canonicalize(&repo).expect("canonical repo root"),
         );
     }
-
-    #[test]
-    fn embedded_web_terminal_copy_shortcut_uses_ctrl_shift_c() {
-        let html = include_str!("../web/index.html");
-
-        assert!(
-            html.contains("function installTerminalCopyHandlers"),
-            "expected web terminal copy handler bootstrap in embedded html",
-        );
-        assert!(
-            html.contains("terminal.attachCustomKeyEventHandler"),
-            "expected xterm custom key handler for copy shortcut",
-        );
-        assert!(
-            html.contains("event.ctrlKey"),
-            "expected Ctrl modifier handling in embedded html",
-        );
-        assert!(
-            html.contains("event.shiftKey"),
-            "expected Shift modifier handling in embedded html",
-        );
-    }
-
-    #[test]
-    fn embedded_web_terminal_drag_selection_copies_on_mouseup() {
-        let html = include_str!("../web/index.html");
-
-        assert!(
-            html.contains("terminalRoot.addEventListener(\"mousedown\""),
-            "expected drag selection tracking in embedded html",
-        );
-        assert!(
-            html.contains("window.addEventListener(\"mouseup\"") && html.contains("handleMouseUp"),
-            "expected mouse release copy handling in embedded html",
-        );
-        assert!(
-            html.contains("function copyTerminalSelection"),
-            "expected clipboard copy helper in embedded html",
-        );
-        assert!(
-            html.contains("navigator.clipboard.writeText"),
-            "expected clipboard write path in embedded html",
-        );
-    }
-
-    #[test]
-    fn embedded_web_terminal_clipboard_fallback_restores_terminal_focus() {
-        let html = include_str!("../web/index.html");
-
-        assert!(
-            html.contains("restoreFocus"),
-            "expected clipboard fallback to invoke restoreFocus after textarea copy",
-        );
-        assert!(
-            html.contains("writeClipboardText(selection")
-                && html.contains("runtime.terminal.focus()"),
-            "expected selection copy path to pass terminal focus restoration",
-        );
-    }
-
-    #[test]
-    fn embedded_web_repo_browser_scroll_surfaces_block_canvas_pan_at_edges() {
-        let html = include_str!("../web/index.html");
-        let scroll_gate = regex::Regex::new(
-            r"if\s*\(\s*!event\.ctrlKey\s*&&\s*!event\.metaKey\s*&&\s*nativeWheelScrollSurface\s*\)\s*\{\s*return;\s*\}",
-        )
-        .expect("valid regex");
-
-        assert!(
-            html.contains("function findNativeWheelScrollSurface"),
-            "expected embedded html to define a repo browser wheel routing helper",
-        );
-        assert!(
-            html.contains(".branch-scroll") && html.contains(".file-tree-scroll"),
-            "expected embedded html to reference repo browser scroll containers",
-        );
-        assert!(
-            html.contains("const nativeWheelScrollSurface = findNativeWheelScrollSurface(event.target);"),
-            "expected plain wheel handling to route repo browser surfaces through the shared helper",
-        );
-        assert!(
-            scroll_gate.is_match(html),
-            "expected plain wheel input on repo browser surfaces to stay within the window even at scroll edges",
-        );
-        assert!(
-            !html.contains("canScrollSurfaceConsumeWheelDelta"),
-            "expected repo browser wheel routing to stop using edge fallback heuristics",
-        );
-    }
 }
 
 fn normalize_active_tab_id(
@@ -2053,7 +1966,7 @@ impl EmbeddedServer {
         let (shutdown_tx, shutdown_rx) = oneshot::channel();
 
         let app = Router::new()
-            .route("/", get(index_handler))
+            .route("/", get(embedded_web::index_handler))
             .route("/healthz", get(health_handler))
             .route("/ws", get(websocket_handler))
             .with_state(ServerState { proxy, clients });
@@ -2082,10 +1995,6 @@ impl EmbeddedServer {
             let _ = tx.send(());
         }
     }
-}
-
-async fn index_handler() -> Html<&'static str> {
-    Html(include_str!("../web/index.html"))
 }
 
 async fn health_handler() -> &'static str {
